+++ hydra-node/examples/Hydra/Events/S3.hs
@@ -23,7 +23,7 @@ import Conduit (
 import Control.Lens (view)
 import Data.Aeson qualified as Aeson
 import Data.List (stripPrefix)
-import Hydra.Events (EventId, EventSink (..), EventSource (..), HasEventId, getEventId)
+import Hydra.Events (EventId, EventSink (..), EventSource (..), HasEventId, getEventId, mkEventSink)
 
 -- | Create a new event source and sink that stores events in AWS S3.
 newS3EventStore :: (HasEventId e, ToJSON e, FromJSON e) => AWS.BucketName -> IO (EventSource e IO, EventSink e IO)
@@ -31,7 +31,7 @@ newS3EventStore bucketName = do
   env <- AWS.newEnv AWS.discover
   pure
     ( EventSource{sourceEvents = sourceEvents env}
-    , EventSink{putEvent = putEvent env}
+    , mkEventSink (putEvent env)
     )
  where
   putEvent env e = do